requirements.txt
anyio==3.0.1
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.5.6
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.1
django-allauth==0.44.0
django-countries==7.2.1
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-formtools==2.3
django-heroku==0.3.1
django-jsonfield==1.4.1
django-otp==1.1.1
django-phonenumber-field==5.2.0
django-two-factor-auth==1.13.1
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.17.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.4.2
fonttools==4.27.1
fpdf==1.7.2
googletrans==3.1.0a0
graphviz==0.16
gunicorn==20.1.0

hstspreload==2021.10.1
html5lib==1.1

image==1.5.33
importlib-metadata==4.0.1
iniconfig==1.1.1
ipykernel==5.5.4
ipython==7.23.1
langcodes==3.3.0
lazy-object-proxy==1.6.0
numpy==1.20.3
oauthlib==3.0.1
opencv-python==4.5.5.62
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.9.1
py==1.11.0
PyPDF2==1.26.0
pyrsistent==0.17.3
pytesseract==0.3.8
pytest==7.0.1
tldextract==3.1.2
virtualenv==20.13.1
virtualenvwrapper-win==1.2.7


